+++ internal/cli/atlas/cloudproviders/cloud_providers.go
@@ -12,16 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cloudprovider
+package cloudproviders
 
 import (
 	"github.com/mongodb/mongocli/internal/cli"
-	"github.com/mongodb/mongocli/internal/cli/atlas/cloudprovider/accessroles"
+	"github.com/mongodb/mongocli/internal/cli/atlas/cloudproviders/accessroles"
 	"github.com/spf13/cobra"
 )
 
 func Builder() *cobra.Command {
-	const use = "cloudProvider"
+	const use = "cloudProviders"
 	cmd := &cobra.Command{
 		Use:     use,
 		Aliases: cli.GenerateAliases(use),
